refactor(gui): Rust A-grade pass on src-tauri
Pure refactor — zero behaviour change visible to React frontend.

Major changes:
- Split 4028-line lib.rs into 15 commands/* modules + thin orchestrator (468 lines)
- Add error.rs (AppError enum, manual Serialize preserves wire format)
- Add state.rs (Tauri-managed AppState, drops 3 static Mutex globals)
- Add paths.rs (canonicalize-checked safe_join + safe_join_relative + safe_new_file, validate_filename + validate_relative_path)
- Add config.rs (serde_yaml structs for read paths)
- Add yaml_writer.rs (comment-preserving line-based writers, with regression tests)
- Fix Box::leak in update_memory_entry (memory leak per edit)
- Hoist hot-path regexes to LazyLock (egress + worktrees, 6 patterns)
- Migrate get_usage_stats to async reqwest::Client
- Replace 50ms hotkey poll loop with notify file watcher
- Path canonicalize + filename validation across notes/agents/memory
- Drop tokio "full" feature (unused), drop reqwest blocking feature
- cargo clippy --all-targets -- -D warnings now passes clean
- Tests: 4 -> 35

Spec: docs/superpowers/specs/2026-05-04-rust-a-grade-refactor-design.md
Plan: docs/superpowers/plans/2026-05-04-rust-a-grade-refactor.md
