+++ gui/src-tauri/Cargo.toml
@@ -17,11 +17,14 @@ tauri = { version = "2", features = ["tray-icon", "image-png"] }
 tauri-plugin-opener = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-tokio = { version = "1", features = ["full", "net"] }
+serde_yaml = "0.9"
+thiserror = "2"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "net"] }
 image = "0.24"
 regex = "1"
 dirs = "5"
 chrono = "0.4"
 uuid = { version = "1", features = ["v4"] }
-reqwest = { version = "0.12", features = ["blocking", "json", "rustls-tls"], default-features = false }
+reqwest = { version = "0.12", features = ["json", "rustls-tls"], default-features = false }
+notify = { version = "7", default-features = false, features = ["macos_kqueue"] }
