+++ gui/src-tauri/src/commands/mod.rs
@@ -1,0 +1,17 @@
+//! Tauri IPC command handlers grouped by domain.
+
+pub mod agents;
+pub mod claude_config;
+pub mod clipboard;
+pub mod github;
+pub mod history;
+pub mod hotkeys;
+pub mod inbox;
+pub mod lifecycle;
+pub mod memory;
+pub mod neuralguard;
+pub mod notes;
+pub mod overlay;
+pub mod remote;
+pub mod usage;
+pub mod worktrees;
